+++ changelog.md
@@ -1,2 +1,4 @@
 * Now dynamic color can be applied to anything
 * Removed duplicated class. Simple clean up and a bunch of renamed
+* Added Molang animation support
+* Added api for player animations